requirements.txt
numpy==2.2.5
pandas==2.2.3
torch==2.6.0
pytorch-lightning==2.5.1.post0
torchmetrics==1.7.1
transformers==4.51.3
datasets==3.5.0
huggingface-hub==0.30.2
wandb==0.19.10
tqdm==4.67.1
requests==2.32.3
PyYAML==6.0.2
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
fsspec==2024.12.0
filelock==3.18.0
dill==0.3.8
multiprocess==0.70.16
xxhash==3.5.0
pyarrow==19.0.1
safetensors==0.5.3
hydra-core==1.3.2
omegaconf==2.3.0
matplotlib==3.10.1
seaborn==0.13.2